requirements.txt
ultralytics
torch
fastapi
uvicorn
python-multipart
opencv-python-headless
pillow
pyyaml
tqdm
numpy
pillow-heif